% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rqualis_functions.R
\name{get_options}
\alias{get_options}
\title{Obter opções dos fórmulários.}
\usage{
get_options(request, form = "evento")
}
\arguments{
\item{request}{Objeto gerado pela função \code{get_sucupira_page}}

\item{form}{Nome do campo do formulário.}
}
\value{
Retorna um \code{tibble} contendo as opções do campo especificado.
}
\description{
Essa função permite que você obtenha uma tabela contendo os
valores de um campo específico.
}
\examples{
request <- get_sucupira_page()
get_options(request)
}
\keyword{opções}
