% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rqualis_functions.R
\name{get_qualis_table}
\alias{get_qualis_table}
\title{Obter tabela de periódicos.}
\usage{
get_qualis_table(qualis_page, cod_event = "14", cod_area = "")
}
\arguments{
\item{qualis_page}{refere-se o objeto gerado pela função \code{get_sucupira_page}}

\item{area}{refere-se a área que se deseja obter a tabela dos periódicos}

\item{event}{refere-se ao evento de classificação dos periódiocos se é do triênio
ou quadriênio}
}
\value{
Retorna um dataframe contendo o valor dos campos de opções do menu informado.
}
\description{
Essa função permite que você obtenha uma tabela contendo os
valores de um campo específico.
}
\examples{

sucupira_pg <- get_sucupira_page()
tb <- get_qualis_table(sucupira_get = sucupira_pg)
print(head(tb))


}
