% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rqualis_functions.R
\name{get_sucupira_page}
\alias{get_sucupira_page}
\title{Obter página para consultar qualis.

Realiza GET na página  do \href{https://sucupira.capes.gov.br/sucupira/public/consultas/coleta/veiculoPublicacaoQualis/listaConsultaGeralPeriodicos.jsf}{Sucupira}}
\usage{
get_sucupira_page()
}
\value{
Retorna a operação obtida pelo GET do rvest no site do sucupira.
}
\description{
Obter página para consultar qualis.

Realiza GET na página  do \href{https://sucupira.capes.gov.br/sucupira/public/consultas/coleta/veiculoPublicacaoQualis/listaConsultaGeralPeriodicos.jsf}{Sucupira}
}
\examples{
sucupira_get <- get_sucupira_page()
print(sucupira_get$headers)
}
