% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rqualis-package.R
\docType{package}
\name{Rqualis-package}
\alias{Rqualis}
\alias{Rqualis-package}
\title{Rqualis: Consulta de qualis na plataforma sucupira.}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Esse pacote permite realizar consultas de qualis de periódicos
    disponíveis na plataforma sucupira.
}
\author{
\strong{Maintainer}: Jessica Cardoso \email{pyrjessica@gmail.com}

}
\keyword{internal}
